Enhanced Metadata Support, SickBeard, CouchPotato
Added support for categories as found in the XMLTV data, lookups and
storage of additional meta data such as the TVDBId, TMDBId, season and
epi

Added a TVDB library
Added a TMDB library
Added a SickBeard library
Added a CouchPotato library
Added columns to the programs table in the sqlite database to capture
seriesId (tvdb), movieId (imdb), episode and season numbers, new flag,
category.
Removed all of the test code sent yesterday that was replacing the
behavior of a button
Added loading of the new program database columns when those database
rows for each program in the source are inserted. (no matter the source
type XMLTV, YOUSEE, etc, all will be able to use these features so long
as the source parser is updated to pull the information from somewhere
in the source data)
Enhanced the XMLTV parser to load the new program attributes (columns)
listed above from the XMLTV data combined with additional sources such
as TVDB & TMDB
Modified the epg viewer code to apply a color to the program based upon
it's category (dark green = sports, light green = news, red = movies,
orange = kids). This actually mirrors verizon fios color coding.
The icons for each program in the EPG were swapped out while changing
the category colors to use a vertical gradient instead of a horizontal
one (stretches better horizontally, and again, matches the verizon fios
interface)
Added "Record with SickBeard" option in the menu for a program. It will
add the show to sickbeard with all prior episodes set to "skipped'
Added an icon to the info panel on the EPG view for showing the
sickbeard logo when the show is managed by sickbeard
Added settings pages for each of the new libraries to configure the base
url's, api keys, and enable flag