Continued Sickbeard Integrations
The program options screen is now functional for scheduling TV Shows to
be managed by Sickbeard. You can also select to record just a single
episode and it will setup the show if it is not already, and mark that
episode as wanted". By default, all newly scheduled shows (series) have
their past episodes set to skipped.